refactor(spark): share ResolveColumnsForInsertInto preprocessing across Spark 4.x
The three-level preprocess logic (partition-spec normalization, user-specified
column projection, output-column resolution with default-value support, and
the arity-mismatch error rewrite) is byte-identical across Spark 4.0-4.2, so
it moves into the abstract HoodieSpark4ResolveColumnsForInsertInto base class
in hudi-spark4-common. The per-version case classes keep only apply(), whose
InsertIntoStatement pattern arity differs in Spark 4.2 (two fields added).

The DataSourceV2ToV1Fallback rules stay fully per-version: every adjacent
pair differs in InsertIntoStatement and/or DataSourceV2Relation case-class
shapes, so no clean seam exists within the family. Spark 3.5 keeps its full
copy of ResolveColumnsForInsertInto since hudi-spark3-common also compiles
against Spark 3.3/3.4, which lack the required APIs. Rule class names are
unchanged (HoodieAnalysis instantiates them reflectively by name).
