+++ hudi-spark-datasource/hudi-spark4.2.x/src/main/scala/org/apache/spark/sql/hudi/analysis/HoodieSpark42Analysis.scala
@@ -20,20 +20,16 @@ package org.apache.spark.sql.hudi.analysis
 import org.apache.hudi.{DefaultSource, EmptyRelation, HoodieBaseRelation}
 import org.apache.hudi.SparkAdapterSupport.sparkAdapter
 
-import org.apache.spark.sql.{AnalysisException, SparkSession}
-import org.apache.spark.sql.catalyst.analysis.{ResolveInsertionBase, TableOutputResolver}
-import org.apache.spark.sql.catalyst.catalog.{CatalogTable, HiveTableRelation}
+import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.catalyst.catalog.HiveTableRelation
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.errors.DataTypeErrors.toSQLId
-import org.apache.spark.sql.errors.QueryCompilationErrors
-import org.apache.spark.sql.execution.datasources.{HadoopFsRelation, LogicalRelation, PreprocessTableInsertion}
+import org.apache.spark.sql.execution.datasources.{HadoopFsRelation, LogicalRelation}
 import org.apache.spark.sql.execution.datasources.v2.DataSourceV2Relation
 import org.apache.spark.sql.hudi.ProvidesHoodieConfig
 import org.apache.spark.sql.hudi.catalog.HoodieInternalV2Table
 import org.apache.spark.sql.sources.InsertableRelation
 import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.util.PartitioningUtils.normalizePartitionSpec
 
 /**
  * NOTE: PLEASE READ CAREFULLY
@@ -74,28 +70,11 @@ case class HoodieSpark42DataSourceV2ToV1Fallback(sparkSession: SparkSession) ext
 }
 
 /**
- * In Spark 3.5, the following Resolution rules are removed,
- * [[ResolveUserSpecifiedColumns]] and [[ResolveDefaultColumns]]
- * (see code changes in [[org.apache.spark.sql.catalyst.analysis.Analyzer]]
- * from https://github.com/apache/spark/pull/41262).
- * The same logic of resolving the user specified columns and default values,
- * which are required for a subset of columns as user specified compared to the table
- * schema to work properly, are deferred to [[PreprocessTableInsertion]] for v1 INSERT.
- *
- * Note that [[HoodieAnalysis]] intercepts the [[InsertIntoStatement]] after Spark's built-in
- * Resolution rules are applies, the logic of resolving the user specified columns and default
- * values may no longer be applied. To make INSERT with a subset of columns specified by user
- * to work, this custom resolution rule [[HoodieSpark42ResolveColumnsForInsertInto]] is added
- * to achieve the same, before converting [[InsertIntoStatement]] into
- * [[InsertIntoHoodieTableCommand]].
- *
- * The implementation is copied and adapted from [[PreprocessTableInsertion]]
- * https://github.com/apache/spark/blob/d061aadf25fd258d2d3e7332a489c9c24a2b5530/sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/rules.scala#L373
- *
- * Also note that, the project logic in [[ResolveImplementationsEarly]] for INSERT is still
- * needed in the case of INSERT with all columns in a different ordering.
+ * Resolution rule resolving the user specified columns and default values of
+ * [[InsertIntoStatement]] for Spark 4.2; see [[HoodieSpark4ResolveColumnsForInsertInto]]
+ * for the shared preprocessing logic and the rationale.
  */
-case class HoodieSpark42ResolveColumnsForInsertInto() extends ResolveInsertionBase {
+case class HoodieSpark42ResolveColumnsForInsertInto() extends HoodieSpark4ResolveColumnsForInsertInto {
   // NOTE: This is copied from [[PreprocessTableInsertion]] with additional handling of Hudi relations
   override def apply(plan: LogicalPlan): LogicalPlan = {
     plan match {
@@ -120,92 +99,6 @@ case class HoodieSpark42ResolveColumnsForInsertInto() extends ResolveInsertionBa
           case _ => i
         }
       case _ => plan
-    }
-  }
-
-  private def preprocess(insert: InsertIntoStatement,
-                         catalogTable: Option[CatalogTable]): InsertIntoStatement = {
-    preprocess(insert, catalogTable, catalogTable.map(_.partitionSchema).getOrElse(new StructType()))
-  }
-
-  private def preprocess(insert: InsertIntoStatement,
-                         catalogTable: Option[CatalogTable],
-                         partitionSchema: StructType): InsertIntoStatement = {
-    val tblName = catalogTable.map(_.identifier.quotedString).getOrElse("unknown")
-    preprocess(insert, tblName, partitionSchema, catalogTable)
-  }
-
-  // NOTE: this is copied from [[PreprocessTableInsertion]] with additional logic
-  // to unset user-specified columns at the end
-  private def preprocess(insert: InsertIntoStatement,
-                         tblName: String,
-                         partColNames: StructType,
-                         catalogTable: Option[CatalogTable]): InsertIntoStatement = {
-
-    val normalizedPartSpec = normalizePartitionSpec(
-      insert.partitionSpec, partColNames, tblName, conf.resolver)
-
-    val staticPartCols = normalizedPartSpec.filter(_._2.isDefined).keySet
-    val expectedColumns = insert.table.output.filterNot(a => staticPartCols.contains(a.name))
-
-    val partitionsTrackedByCatalog = catalogTable.isDefined &&
-      catalogTable.get.partitionColumnNames.nonEmpty &&
-      catalogTable.get.tracksPartitionsInCatalog
-    if (partitionsTrackedByCatalog && normalizedPartSpec.nonEmpty) {
-      // empty partition column value
-      if (normalizedPartSpec.values.flatten.exists(v => v != null && v.isEmpty)) {
-        val spec = normalizedPartSpec.map(p => p._1 + "=" + p._2).mkString("[", ", ", "]")
-        throw QueryCompilationErrors.invalidPartitionSpecError(
-          s"The spec ($spec) contains an empty partition column value")
-      }
-    }
-
-    // Create a project if this INSERT has a user-specified column list.
-    val hasColumnList = insert.userSpecifiedCols.nonEmpty
-    val query = if (hasColumnList) {
-      createProjectForByNameQuery(tblName, insert)
-    } else {
-      insert.query
-    }
-    val newQuery = try {
-      TableOutputResolver.resolveOutputColumns(
-        tblName,
-        expectedColumns,
-        query,
-        byName = hasColumnList || insert.byName,
-        conf,
-        supportColDefaultValue = true)
-    } catch {
-      case e: AnalysisException if staticPartCols.nonEmpty &&
-        (e.getErrorClass == "INSERT_COLUMN_ARITY_MISMATCH.NOT_ENOUGH_DATA_COLUMNS" ||
-          e.getErrorClass == "INSERT_COLUMN_ARITY_MISMATCH.TOO_MANY_DATA_COLUMNS") =>
-        val newException = e.copy(
-          errorClass = Some("INSERT_PARTITION_COLUMN_ARITY_MISMATCH"),
-          messageParameters = e.messageParameters ++ Map(
-            "tableColumns" -> insert.table.output.map(c => toSQLId(c.name)).mkString(", "),
-            "staticPartCols" -> staticPartCols.toSeq.sorted.map(c => toSQLId(c)).mkString(", ")
-          ))
-        newException.setStackTrace(e.getStackTrace)
-        throw newException
-    }
-    if (normalizedPartSpec.nonEmpty) {
-      if (normalizedPartSpec.size != partColNames.length) {
-        throw QueryCompilationErrors.requestedPartitionsMismatchTablePartitionsError(
-          tblName, normalizedPartSpec, partColNames)
-      }
-
-      // NOTE: Hudi converts [[InsertIntoStatement]] to [[InsertIntoHoodieTableCommand]]
-      // and the user specified is no longer need after resolution
-      // (`userSpecifiedCols = Seq()`)
-      insert.copy(query = newQuery, partitionSpec = normalizedPartSpec, userSpecifiedCols = Seq())
-    } else {
-      // All partition columns are dynamic because the InsertIntoTable command does
-      // not explicitly specify partitioning columns.
-      // NOTE: Hudi converts [[InsertIntoStatement]] to [[InsertIntoHoodieTableCommand]]
-      // and the user specified is no longer need after resolution
-      // (`userSpecifiedCols = Seq()`)
-      insert.copy(query = newQuery, partitionSpec = partColNames.map(_.name).map(_ -> None).toMap,
-        userSpecifiedCols = Seq())
     }
   }
 }
